requirements.txt
langgraph>=0.2.55
langchain-core>=0.1.18
langchain-community>=0.3.9
tavily-python>=0.5.0
langchain-ollama>=0.2.1
duckduckgo-search>=7.3.0
beautifulsoup4>=4.13.3
gradio>=4.0.0
requests>=2.31.0
typing-extensions>=4.8.0
python-dotenv>=1.0.0
